Renamed a Python test script
